Initial commit of memo and heuristic based code
This solution uses heuristics to guide the matching process, it does not guarantee an optimal solution. However, it aims to find a reasonably good solution within a reasonable amount of time, which is often desirable in practical situations.

Potential to use optimization methods, modeling soft and hard constraints to improve the optimal team formation and scalability as number of sign ups increase.